% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{collin_pairtest}
\alias{collin_pairtest}
\title{Test for potential collinearity issues in a multivariable Cox model via paired comparisons}
\usage{
collin_pairtest(fit, digits = 4, ...)
}
\description{
Test for potential collinearity issues in a multivariable Cox model via paired comparisons
}
