% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmodels.R
\name{coxambi.coxph}
\alias{coxambi.coxph}
\title{Wrapper for a readily fitted Cox model and perform single-variable fits per each fitted covariate}
\usage{
\method{coxambi}{coxph}(obj, ...)
}
\description{
Wrapper for a readily fitted Cox model and perform single-variable fits per each fitted covariate
}
