% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmodels.R
\name{coxambi.default}
\alias{coxambi.default}
\title{Not a suitable class provided as first parameter for coxambi}
\usage{
\method{coxambi}{default}(obj, ...)
}
\description{
Not a suitable class provided as first parameter for coxambi
}
