% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmodels.R
\name{coxambi.formula}
\alias{coxambi.formula}
\title{Wrapper for a formula + data formulation of Cox model}
\usage{
\method{coxambi}{formula}(obj, data, ...)
}
\description{
Wrapper for a formula + data formulation of Cox model
}
