% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmodels.R
\name{coxtab}
\alias{coxtab}
\title{Tabulate multivariable and single variable Cox fit results next to each other}
\usage{
coxtab(obj, data, format, digits = c(coef = 3, p = 3), addN = TRUE, ...)
}
\description{
Tabulate multivariable and single variable Cox fit results next to each other
}
