% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmodels.R
\name{coxambi}
\alias{coxambi}
\title{Fit multivariable Cox ph model for all covariates together with univariable model once per each covariate}
\usage{
coxambi(formula, data, ...)
}
\description{
Fit multivariable Cox ph model for all covariates together with univariable model once per each covariate
}
